Refactor shared packages into util directory
