+++ cmd/kops/root.go
@@ -10,8 +10,8 @@ import (
 	"github.com/spf13/viper"
 	"k8s.io/kops/upup/pkg/api"
 	"k8s.io/kops/upup/pkg/fi"
-	"k8s.io/kops/upup/pkg/fi/vfs"
 	"k8s.io/kops/upup/pkg/kutil"
+	"k8s.io/kops/util/pkg/vfs"
 )
 
 type RootCmd struct {
