+++ upup/pkg/fi/assetstore.go
@@ -4,8 +4,8 @@ import (
 	"fmt"
 	"github.com/golang/glog"
 	"io"
-	"k8s.io/kops/upup/pkg/fi/hashing"
 	"k8s.io/kops/upup/pkg/fi/utils"
+	"k8s.io/kops/util/pkg/hashing"
 	"net/http"
 	"os"
 	"os/exec"
