+++ upup/pkg/fi/http.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 	"github.com/golang/glog"
 	"io"
-	"k8s.io/kops/upup/pkg/fi/hashing"
+	"k8s.io/kops/util/pkg/hashing"
 	"net/http"
 	"os"
 	"path"
