+++ upup/pkg/fi/loader/tree_walker.go
@@ -3,7 +3,7 @@ package loader
 import (
 	"fmt"
 	"github.com/golang/glog"
-	"k8s.io/kops/upup/pkg/fi/vfs"
+	"k8s.io/kops/util/pkg/vfs"
 	"os"
 	"path"
 	"strings"
