+++ upup/pkg/kutil/kubectl.go
@@ -65,7 +65,7 @@ func (k *Kubectl) execKubectl(args ...string) (string, error) {
 	glog.V(2).Infof("Running command: %s", human)
 	output, err := cmd.CombinedOutput()
 	if err != nil {
-		glog.Info("error running %s:", human)
+		glog.Infof("error running %s:", human)
 		glog.Info(string(output))
 		return string(output), fmt.Errorf("error running kubectl")
 	}
