+++ util/pkg/hashing/hash.go
